requirements.txt
# Core dependencies
asgiref==3.8.1
psutil==7.0.0
humanize==4.12.3
certifi==2025.1.31
psycopg2-binary==2.9.9
whitenoise==6.6.0
charset-normalizer==3.4.2
Django==4.2.11  # Downgraded to be compatible with django-celery-beat
python-dotenv==1.0.1  # For loading .env files
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0

# Django REST framework
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-storages==1.14.2

# Celery and related
celery==5.3.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
redis==5.0.3
kombu==5.3.6
billiard==4.2.1
vine==5.1.0
